# openapi-spec.yaml
openapi: 3.0.3
info:
  title: Manus AI API
  description: |
    A comprehensive, versatile AI system built with NestJS that provides specialized AI agents capable of performing various tasks through real AI integration (OpenAI + LangChain), background processing, and system automation.

    ## Key Features
    - **9 Specialized AI Agents** with real AI integration (no mock data)
    - **Background Task Processing** with BullMQ and Redis queues
    - **Session Persistence** for offline capability and long-running tasks
    - **Real-time AI Interactions** with streaming responses via LangChain
    - **System Tool Integration** (file operations, browser automation, shell commands)

    ## Technologies Used
    - **OpenAI + LangChain**: Real AI integration with proper streaming and tool binding
    - **NestJS**: Progressive Node.js framework
    - **BullMQ**: Advanced Redis-based queue system
    - **MongoDB**: NoSQL database with Prisma ORM
    - **TypeScript**: Strongly typed development

  version: 1.0.0
  contact:
    name: Manus AI Team
    email: support@manus.im
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.manus.im/api/v1
    description: Production server
paths:
...
  /session/{sessionId}:
    get:
      summary: Get session status
      description: Get the current status of a session
      tags:
        - Session
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Session ID
          schema:
            type: string
      responses:
        '200':
          description: Session status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
  /chat:
    post:
      summary: Background chat processing
      description: Send a message for background processing with queue management
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackgroundChatRequest'
      responses:
        '202':
          description: Chat task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
  /generate/text:
    post:
      summary: Generate text content
      description: Generate text content using AI with LangChain integration
      tags:
        - Generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextGenerateRequest'
      responses:
        '202':
          description: Text generation task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /generate/code:
    post:
      summary: Generate code
      description: Generate code using AI with specific language and framework parameters
      tags:
        - Generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGenerateRequest'
      responses:
        '202':
          description: Code generation task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /generate/image:
    post:
      summary: Generate image description
      description: Generate image descriptions or concepts using AI
      tags:
        - Generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageGenerateRequest'
      responses:
        '202':
          description: Image generation task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
  /generate/task/{taskId}/retry:
    post:
      summary: Retry failed generation task
      description: Retry a failed generation task
      tags:
        - Generation
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
      responses:
        '200':
          description: Task retried
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /generate/task/{taskId}/cancel:
    post:
      summary: Cancel generation task
      description: Cancel a running generation task
      tags:
        - Generation
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
      responses:
        '200':
          description: Task cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /generate/task/{taskId}:
    delete:
      summary: Delete generation task
      description: Delete a completed or failed generation task
      tags:
        - Generation
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /generate/history/{sessionId}:
    get:
      summary: Get generation history
      description: Get the generation history for a session
      tags:
        - Generation
        - History
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Session ID
          schema:
            type: string
      responses:
        '200':
          description: Generation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GenerationHistoryItem'
...
  /edit:
    post:
      summary: General file editing
      description: Perform general file editing operations using AI
      tags:
        - Edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditRequest'
      responses:
        '202':
          description: Edit task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
components:
  schemas:
...
    SessionStatus:
      type: object
      properties:
        sessionId:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        tasksInQueue:
          type: integer
          description: Number of tasks in queue
        completedTasks:
          type: integer
          description: Number of completed tasks
        configuration:
          type: object
          properties:
            aiEnabled:
              type: boolean
            queueEnabled:
              type: boolean
            maxConcurrentTasks:
              type: integer
    BackgroundChatRequest:
      type: object
      required:
        - message
        - sessionId
      properties:
        message:
          type: string
          description: Message for background processing
        sessionId:
          type: string
          description: Session identifier
        agentType:
          type: string
          description: Type of agent to use
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
    TaskResponse:
      type: object
      properties:
        taskId:
          type: string
          description: Unique task identifier
        status:
          type: string
          enum: [queued, processing, completed, failed]
        estimatedTime:
          type: string
          description: Estimated completion time
        sessionId:
          type: string
...
    TextGenerateRequest:
      type: object
      required:
        - prompt
        - sessionId
      properties:
        prompt:
          type: string
          description: Text generation prompt
        sessionId:
          type: string
        parameters:
          type: object
          properties:
            maxTokens:
              type: integer
              default: 2000
            temperature:
              type: number
              format: float
              default: 0.7
            style:
              type: string
              description: Writing style preference
    CodeGenerateRequest:
      type: object
      required:
        - prompt
        - sessionId
      properties:
        prompt:
          type: string
          description: Code generation prompt
        language:
          type: string
          description: Programming language
        framework:
          type: string
          description: Framework or library to use
        sessionId:
          type: string
    ImageGenerateRequest:
      type: object
      required:
        - prompt
        - sessionId
      properties:
        prompt:
          type: string
          description: Image description prompt
        sessionId:
          type: string
    EditRequest:
      type: object
      required:
        - instruction
        - sessionId
      properties:
        instruction:
          type: string
          description: Edit instruction
        target:
          type: string
          description: Target file or content
        sessionId:
          type: string
...
    GenerationHistoryItem:
      type: object
      properties:
        id:
          type: string
        prompt:
          type: string
        result:
          type: string
        type:
          type: string
          enum: [text, code, image]
        timestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [completed, failed]
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...